Simplify types package
