+++ packages/themes/package.json
@@ -42,9 +42,7 @@
     "watch": "tsc -p tsconfig.json --watch",
     "lint": "eslint -c '../../.eslintrc' --fix './src/**/*.ts'"
   },
-  "dependencies": {
-    "@web3modal/types": "^1.0.0"
-  },
+  "dependencies": {},
   "devDependencies": {
     "@babel/cli": "^7.12.1",
     "@babel/core": "^7.12.3",
@@ -59,6 +57,7 @@
     "@types/node": "^14.14.2",
     "@typescript-eslint/eslint-plugin": "^2.24.0",
     "@typescript-eslint/parser": "^2.24.0",
+    "@web3modal/types": "^1.0.0",
     "chai": "^4.2.0",
     "core-js": "^3.6.5",
     "eslint": "^5.16.0",
