--- packages/types/.babelrc
@@ -1,12 +1,0 @@
-{
-  "presets": [
-    [
-      "@babel/preset-env",
-      {
-        "useBuiltIns": "usage",
-        "corejs": 3
-      }
-    ],
-    "@babel/typescript"
-  ]
-}
