--- packages/types/.npmignore
@@ -1,26 +1,0 @@
-*.log
-npm-debug.log*
-
-# Coverage directory used by tools like istanbul
-coverage
-.nyc_output
-
-# Dependency directories
-node_modules
-
-# npm package lock
-package-lock.json
-yarn.lock
-
-# project files
-src
-test
-examples
-CHANGELOG.md
-.travis.yml
-.editorconfig
-.eslintignore
-.eslintrc
-.babelrc
-.gitignore
-.watchmanconfig
