--- packages/types/test/index.spec.js 
@@ -1,7 +1,0 @@
-import "mocha";
-import * as chai from "chai";
-describe("@web3modal/types", () => {
-  it("needs tests", () => {
-    // needs tests
-  });
-});
