--- packages/types/webpack.config.js
@@ -1,16 +1,0 @@
-const path = require("path");
-
-module.exports = {
-  mode: "production",
-  entry: {
-    index: path.resolve(__dirname, "dist", "cjs", "index.js"),
-  },
-  output: {
-    path: path.resolve(__dirname, "dist", "umd"),
-    filename: "[name].min.js",
-    libraryTarget: "umd",
-    library: "Web3ModalTypes",
-    umdNamedDefine: true,
-    globalObject: "this",
-  },
-};
