+++ packages/utils/package.json
@@ -46,8 +46,7 @@
     "detect-browser": "^5.1.0",
     "@web3modal/chain-utils": "^1.0.0",
     "@web3modal/providers": "^1.0.0",
-    "@web3modal/themes": "^1.0.0",
-    "@web3modal/types": "^1.0.0"
+    "@web3modal/themes": "^1.0.0"
   },
   "devDependencies": {
     "@babel/cli": "^7.12.1",
@@ -64,6 +63,7 @@
     "@types/web3": "^1.2.2",
     "@typescript-eslint/eslint-plugin": "^2.24.0",
     "@typescript-eslint/parser": "^2.24.0",
+    "@web3modal/types": "^1.0.0",
     "chai": "^4.2.0",
     "core-js": "^3.6.5",
     "eslint": "^5.16.0",
